ai-services/speech/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
numpy
torch
ffmpeg-python
